Add options.d file for the jvm.options
This commit adds templated config for the rally jvm.options into a
jvm.options.d folder, and turns the top level jvm.options into the same
file as a standard 8.0 deployment would have. The 8.0 file was pulled
from a built snapshot of Elasticsearch.

The biggest changes are the removal of some of the default stanzas that
were already present in the jvm.options that Elasticsearch provides, as
well as nullifying the logging.

Relates #43